Update BayesMCToyExamples.py
Add code for fair and an approximation of unfair dies. Built a case where you pull a random die from a bag of 6 dies where one of the 6 always rolls a 6 but the others are fair. If you pull a die an roll it twice both rolling a 6 what is the probability that you pulled the unfair die?